Added transactions!

+++ server/src/models/transaction.rs
@@ -1,6 +1,9 @@
 use chrono::Utc;
-use serde::{Deserialize};
+use diesel::{Queryable, Insertable};
+use serde::{Deserialize, Serialize};
 use std::fmt;
+
+use crate::schema::transactions;
 
 #[derive(PartialEq, Clone, Copy)]
 pub enum TransactionKind {
@@ -21,18 +24,19 @@ impl fmt::Display for TransactionKind {
     }
 }
 
-#[derive(Deserialize)]
+//NOTE: Mutations are in cents and reflect the effect of the transaction on the account balance. This is done so with a simple sql query we can check the current balance of an account by summing all mutations for that account.
+#[derive(Serialize, Deserialize, Queryable, Insertable)]
 pub struct Transaction {
     pub id: i32,
 	pub account: String,
 	pub created_at: String,
     pub kind: String,
-	pub mutation: f32,
+	pub mutation: i32,
     pub recipient: String
 }
 
 impl Transaction {
-	pub fn new(account: String, transaction_kind: TransactionKind, mutation: f32, recipient: String) -> Self {
+	pub fn new(account: String, transaction_kind: TransactionKind, mutation: i32, recipient: String) -> Self {
 		Self {
             id: 0,
             account,
